Adding regex support for matrix run logs
This implementation supports the matching of regular expressions
contained in the logs of single matrix runs. So far, with the
"regex"-option the plugin only scans the parent job's build log. Usually
this log contains little to no information about the actual problems
that occurred during the individual matrix combination runs. My approach
checks each log of the matrix job runs and rebuilds if the regex is
found in any of them. The implementation also works fine with the
"rebuild only failed parts of the matrix"-option.