[JENKINS-59222] document user on retry
Adding a UserIdCause to the list of causes. Avoid creating a duplicate
CauseAction. Filter out previous users and naginator causes to avoid
lengthy cause lists.
The changed way of handling the cause requires to add a
BuildTriggerBadgeDeactivator otherwise we will get 2 badges.
Store the cause in the action so we can show the cause as tooltip (this
includes the build number).